services/atlas-runtime/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
Authlib==1.7.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
cryptography==48.0.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
frozenlist==1.8.0
google-adk==1.33.0
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
google-cloud-aiplatform==1.151.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.38.0
google-cloud-bigtable==2.38.0
google-cloud-core==2.6.0
google-cloud-dataplex==2.19.0
google-cloud-discoveryengine==0.13.12
google-cloud-firestore==2.27.0
google-cloud-iam==2.23.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.38.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.28.0
google-cloud-spanner==3.66.0
google-cloud-speech==2.39.0
google-cloud-storage==3.10.1
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.75.0
google-resumable-media==2.9.0
googleapis-common-protos==1.75.0
graphviz==0.21
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
importlib_metadata==8.7.1
joserfc==1.6.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Mako==1.3.12
MarkupSafe==3.0.3
mcp==1.27.1
mmh3==5.2.1
multidict==6.7.1
opentelemetry-api==1.41.1
opentelemetry-exporter-gcp-logging==1.12.0a0
opentelemetry-exporter-gcp-monitoring==1.12.0a0
opentelemetry-exporter-gcp-trace==1.12.0
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-http==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-resourcedetector-gcp==1.12.0a0
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
packaging==26.2
propcache==0.5.2
proto-plus==1.28.0
protobuf==6.33.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
PyJWT==2.12.1
pyOpenSSL==26.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlalchemy-spanner==1.18.0
sqlparse==0.5.5
sse-starlette==3.4.2
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.7.0
uvicorn==0.46.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.23.0
zipp==3.23.1
